requirements.txt
fastapi~=0.115.12
uvicorn
python-dotenv~=1.1.0
pydantic>=2
pydantic-settings~=2.9.1
motor~=3.7.1
httpx
requests~=2.31.0
bcrypt~=4.3.0
pymongo~=4.13.0
PyMsgBox~=1.0.9